NET-7813 - DNS : SERVFAIL when resolving PTR records
